+++ python/cupy/samples/cuda.py
@@ -1,0 +1,5 @@
+#!/bin/env python
+
+import cupy as cp
+
+print("Is GPU[cuda] available? ->", cp.cuda.is_available())